refactor: useSetupSteps 훅 분리
