+++ src/featured/interview/components/InterviewSetupModal.tsx
@@ -3,6 +3,7 @@
 import { AnimatePresence, motion } from 'framer-motion'
 import { CheckCircle2, ChevronRight } from 'lucide-react'
 import { useCallback, useEffect, useRef, useState } from 'react'
+import { useSetupSteps } from '@/featured/interview/hooks/useSetupSteps'
 import { CameraStep } from '@/featured/interview/components/setup/CameraStep'
 import { DocumentStep } from '@/featured/interview/components/setup/DocumentStep'
 import { MicStep } from '@/featured/interview/components/setup/MicStep'
@@ -12,7 +13,7 @@ import { useCameraHandler } from '@/featured/interview/hooks/useCameraHandler'
 import type { useInterview } from '@/featured/interview/hooks/useInterview'
 import { useMicPermission } from '@/featured/interview/hooks/useMicPermission'
 import { useMicRecorder } from '@/featured/interview/hooks/useMicRecorder'
-import { type InterviewFileType, type StepStatus } from '@/featured/interview/types'
+import { type InterviewFileType } from '@/featured/interview/types'
 import { Button } from '@/shared/ui/button'
 import { Dialog, DialogContent, DialogTitle } from '@/shared/ui/dialog'
 import {
@@ -38,12 +39,17 @@ const RESUME_LOCKED_STEPS = [1, 2]
 const STEP2_LOCKED = [2]
 
 export function InterviewSetupModal({ session, isRestart = false }: InterviewSetupModalProps) {
-  const [completedSteps, setCompletedSteps] = useState<Set<number>>(() =>
-    isRestart ? new Set([1, 2]) : new Set(),
-  )
-  const [isStep2Locked, setIsStep2Locked] = useState(isRestart)
-  const [currentStep, setCurrentStep] = useState(() => (isRestart ? 3 : 1))
-  const [maxReachedStep, setMaxReachedStep] = useState(() => (isRestart ? 3 : 1))
+  const {
+    isStep2Locked,
+    setIsStep2Locked,
+    currentStep,
+    maxReachedStep,
+    completeStep,
+    navigateToStep,
+    doneCount,
+    allDone,
+    statuses,
+  } = useSetupSteps(isRestart)
   const [title, setTitle] = useState('')
   const [resume, setResume] = useState<File | null>(null)
   const [portfolio, setPortfolio] = useState<File | null>(null)
@@ -66,9 +72,7 @@ export function InterviewSetupModal({ session, isRestart = false }: InterviewSet
     handlePollingComplete,
   )
   const activeQuestions =
-    isRestart && session.unansweredQuestions?.length
-      ? session.unansweredQuestions
-      : questions
+    isRestart && session.unansweredQuestions?.length ? session.unansweredQuestions : questions
 
   const isQuestionsReady = !!activeQuestions && activeQuestions.length > 0
 
@@ -98,15 +102,6 @@ export function InterviewSetupModal({ session, isRestart = false }: InterviewSet
       window.removeEventListener('popstate', handleHistoryBack)
     }
   }, [cleanupSetupDevices, session.showSetup])
-
-  const statuses: StepStatus[] = [1, 2, 3, 4].map((step) => {
-    if (step === currentStep) return 'active'
-    if (completedSteps.has(step)) return 'done'
-    return 'pending'
-  })
-
-  const doneCount = completedSteps.size
-  const allDone = doneCount === 4
 
   const handleDocumentConfirm = async () => {
     if (!session.interviewId || !resume) return
@@ -184,71 +179,32 @@ export function InterviewSetupModal({ session, isRestart = false }: InterviewSet
     }
   }
 
-  const completeStep = (step: number) => {
-    const newCompleted = new Set([...completedSteps, step])
-    setCompletedSteps(newCompleted)
-    const nextStep = [1, 2, 3, 4].find((s) => !newCompleted.has(s))
-    const dest = nextStep ?? 5
-    setCurrentStep(dest)
-    if (dest > maxReachedStep) setMaxReachedStep(dest)
-  }
-
   const syncInterviewTitle = async () => {
-    if (!session.interviewId) return
-
-    const titleRegex = /^[가-힣a-zA-Z0-9 ]{1,20}$/
-    const nextTitle = title.trim()
-
-    if (!nextTitle || !titleRegex.test(nextTitle)) return
-
-    const result = await updateInterviewTitleAction(session.interviewId, nextTitle)
+    const result = await updateInterviewTitleAction(session.interviewId, title)
     if (!result.success) {
       toast.error(result.message || '면접 제목 수정 실패')
     } else {
-      setTitle(nextTitle)
-    }
-  }
-
-  const navigateToStep = (step: number) => {
-    if (step === 2 && isStep2Locked) return
-    if (step === 1 && isRestart) return
-    if (maxReachedStep >= 4 || completedSteps.has(step)) {
-      if (step === 1) {
-        void syncInterviewTitle()
-      }
-      setCurrentStep(step)
+      setTitle(title)
     }
   }
 
   const handleTitleConfirm = async () => {
-    const titleRegex = /^[가-힣a-zA-Z0-9 ]{1,20}$/
-    const targetTitle = title.trim()
-
-    if (!targetTitle) {
-      toast.error('면접 제목을 입력해주세요.')
-      return
-    }
-    if (!titleRegex.test(targetTitle)) {
-      toast.error('제목은 공백을 포함한 한글, 영어, 숫자 1~20자만 가능합니다.')
-      return
-    }
-
     if (session.interviewId) {
-      const result = await updateInterviewTitleAction(session.interviewId, targetTitle)
+      const result = await updateInterviewTitleAction(session.interviewId, title)
       if (result.success) {
-        setTitle(targetTitle)
+        setTitle(title)
         trackEvent('complete_title_input', { category: 'interview_setup' })
         completeStep(1)
       } else {
         toast.error(result.message || '면접 제목 수정 실패')
       }
     } else {
-      const result = await createInterviewAction(targetTitle)
+      const result = await createInterviewAction(title)
       if (result.success) {
         if (result.data) {
           session.setInterviewId(result.data.intvId)
         }
-        setTitle(targetTitle)
+        setTitle(title)
         trackEvent('complete_title_input', { category: 'interview_setup' })
         completeStep(1)
       } else {
@@ -357,7 +313,7 @@ export function InterviewSetupModal({ session, isRestart = false }: InterviewSet
           <SetupSidebar
             statuses={statuses}
             doneCount={doneCount}
-            onStepClick={navigateToStep}
+            onStepClick={(step) => navigateToStep(step, () => void syncInterviewTitle())}
             freeNavigation={maxReachedStep >= 4}
             lockedSteps={isRestart ? RESUME_LOCKED_STEPS : isStep2Locked ? STEP2_LOCKED : undefined}
           />
